✨ feat: Add modals for bookmarks, notifications, logout, and post details in profile page
